requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
imageio==2.37.0
imageio-ffmpeg==0.6.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
psutil==6.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
